updated kadane's algo
